requirements.txt
Flask
zai-sdk
flask-cors
python-dotenv
flask-sqlalchemy
psycopg2-binary